Refactor: PR #350 restart with zero waste abstraction cleanup
- Decommission legacy `useSearchHighlight.tsx` and integrate logic inline in `GlobalSearch.tsx`.
- Prune unused `--color-accent-brand` variables and over-engineered `scanline` keyframes from `src/index.css`.
- Standardize environment variables via new unified `scripts/utils/env.mjs` for capture scripts (`audit_capture.py` and `ai-debugger.mjs`).
- Migrate root `.js` configuration scripts to `.mjs` to enforce consistent module resolution.
