+++ src/components/GlobalSearch.tsx
@@ -1,8 +1,8 @@
 import { Search, X, Hash, CornerDownLeft, Sparkles } from 'lucide-react';
 import { Box, Stack, Text } from '@/layouts/Primitives';
 import { useGlobalSearch } from '@/hooks/useGlobalSearch';
-import { useSearchHighlight } from '@/hooks/useSearchHighlight';
-import { useRef, MouseEvent, ChangeEvent } from 'react';
+import { getHighlightedParts } from '@/lib/utils';
+import { useRef, MouseEvent, ChangeEvent, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useHotkeys, useCommandKey } from '@/hooks/useHotkeys';
 
@@ -15,8 +15,18 @@ interface SearchResult {
 
 export function GlobalSearch() {
   const { query, setQuery, results, isOpen, open, close } = useGlobalSearch();
-  const { highlight } = useSearchHighlight(query);
   const inputRef = useRef<HTMLInputElement>(null);
+
+  const highlight = useCallback((text: string) => {
+    const parts = getHighlightedParts(text, query);
+    if (parts.length === 1) return text;
+
+    return parts.map((part, i) =>
+      part.toLowerCase() === query.toLowerCase()
+        ? <Box as="span" key={i} radius="industrial" paddingX={0.5} surface="accent" weight="font-bold">{part}</Box>
+        : part
+    );
+  }, [query]);
   const navigate = useNavigate();
 
   // 1. The Context Reset: Close on route change
