+++ src/styles/tokens.css
@@ -7,7 +7,6 @@
   --raw-color-accent: #007BFF;
   --raw-color-accent-shadow: rgba(255, 127, 80, 0.3);
   --raw-color-accent-navy: #1A2B3C;
-  --raw-color-accent-brand: #007BFF;
   --raw-color-text-main: #1A2B3C;
   --raw-color-text-body: #2D3748;
   --raw-color-text-dim: #4B4B4B;
