+++ tsconfig.node.json
@@ -7,5 +7,5 @@
     "noEmit": true,
     "strict": true
   },
-  "include": ["vite.config.ts", "postcss.config.js", "tailwind.config.js", "playwright.config.ts", "scripts/content-loader.ts"]
+  "include": ["vite.config.ts", "postcss.config.mjs", "tailwind.config.mjs", "playwright.config.ts", "scripts/content-loader.ts"]
 }
